requirements.txt
requests
pymysql
python-dotenv
pandas
streamlit
